game adjustments
